/@/store/domain/staredChannelsの分離
